Modules to customize the appearence of a phylogeny

show_modules()

Show a data frame of the available Shiny modules.

srv_list/uis_list

Simple named lists with the available server and ui components. These are used internally in make_phynotate to generate the application UI and server.

layout_ui()/layout_server()

Change the layout (rectangular, slanted, circular) and directon (left, right, ...) of a phylogeny displayed through the draw module. See ?make_phynotate for example usage.

branches_ui()/branches_server()

Change the appearence of the branches (line color, weight) of a phylogeny displayed through the draw module. See ?make_phynotate for example usage.

tips_ui()/tips_server()

Enable and change the appearence of the tip labels (font color, font size, offset) of a phylogeny displayed through the draw module. See ?make_phynotate for example usage.

plotarea_ui()/plotarea_server()

Change the plotting areaof a phylogeny displayed through the draw module. For now, only plot height and width can be modified. See ?make_phynotate for example usage.

misc_ui()/misc_server()

Less common customizations of a phylogeny displayed through the draw module. For example, the to modify the amount of space (degrees) between the first and last tip of a phylogeny plotted in circular layout. See ?make_phynotate for example usage.

show_modules()

srv_list

uis_list

layout_ui(id, phy_data)

layout_server(id, phy_data)

branches_ui(id, phy_data)

branches_server(id, phy_data)

tips_ui(id, phy_data)

tips_server(id, phy_data)

plotarea_ui(id, phy_data)

plotarea_server(id, phy_data)

misc_ui(id, phy_data)

misc_server(id, phy_data)

Arguments

id
namespace id

phy_data
the treedata or phylo object being plotted. This is useful to extract associated data that can be used to map variables onto tree aesthetics.

Format

An object of class list of length 5.

An object of class list of length 5.